+++ CHANGELOG.md
@@ -4,9 +4,11 @@
 * `0.1.x` Releases - [0.1.0](#010) | [0.1.1](#011) | [0.1.2](#012) | [0.1.3](#013) | [0.1.4](#014) | [0.1.5](#015) | [0.1.6](#016) | [0.1.7](#017) | [0.1.8](#018) | [0.1.9](#019)
 
 ## 0.2.2
+Released 11 Nov 2017
 
 #### Features
 * `EMP-10601` Added fetch of specific carousels by id through `FetchCarouselItem`
+*`EMP-10609` Analytics dispatch for download events.
 
 #### Changes
 *`EMP-10242` AnalyticsProvider now supplied through a generator to accomodate association with `Player.MediaAsset`.